pyproject.toml
[project]
name = "pegel-online"
version = "0.1.0"
description = "Adapter for fetching water level data from Pegel Online API and publishing to MQTT"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gcmb-publisher>=0.4.0",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
    "requests>=2.31.0",
]
